train SVM

+++ featureExtraction.h
@@ -6,7 +6,7 @@
 #include <opencv2/imgproc/imgproc.hpp>
 
 double*** compute3DTemplate(double*** hog, const std::vector<int> &dims, int grid_pos_x, int grid_pos_y);
-double* compute1DTemplate(double*** hog, const std::vector<int> &dims, int grid_pos_x, int grid_pos_y, int scale);
+float* compute1DTemplate(double*** hog, const std::vector<int> &dims, int grid_pos_x, int grid_pos_y, int scale);
 void multiscale(cv::Mat img);
 
 #endif
